added more 1-learning
built the calculate q value function. put in the part of the solve mdp for the q value. made a function to set the values and optimal moves for each state based off of q values after running q learning. still not working so I have some questions for prof majercik